[P24-63] feat: [BE] movingPlanId 검증을 위한 validator 구현 및 적용 (#54)
* [P24-63] feat: movingPlanId 검증을 위한 Validator 구현

- 확인을 위해 MovingPlanController에 적용

* [P24-63] refactor: main 브랜치 코드 병합

* [P24-63] feat: 각 controller에 movingPlanId 검증 적용

+++ BE/src/main/java/com/example/p24zip/domain/house/controller/HouseController.java
@@ -9,10 +9,13 @@ import com.example.p24zip.domain.house.dto.response.ChangeHouseNicknameResponseD
 import com.example.p24zip.domain.house.dto.response.GetHouseDetailsResponseDto;
 import com.example.p24zip.domain.house.dto.response.HouseListResponseDto;
 import com.example.p24zip.domain.house.service.HouseService;
+import com.example.p24zip.domain.user.entity.User;
 import com.example.p24zip.global.response.ApiResponse;
+import com.example.p24zip.global.validator.MovingPlanValidator;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -27,43 +30,56 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/plans/{movingPlanId}/houses")
 public class HouseController {
     private final HouseService houseService;
+    private final MovingPlanValidator movingPlanValidator;
 
     @PostMapping
-    public ResponseEntity<ApiResponse<AddHouseResponseDto>> postHouse(@PathVariable Long movingPlanId, @RequestBody @Valid AddHouseRequestDto requestDto){
+    public ResponseEntity<ApiResponse<AddHouseResponseDto>> postHouse(@PathVariable Long movingPlanId, @RequestBody @Valid AddHouseRequestDto requestDto, @AuthenticationPrincipal User user){
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         return ResponseEntity.ok(
             ApiResponse.ok("CREATED", "집 생성에 성공했습니다.", houseService.postHouse(movingPlanId, requestDto))
         );
     }
 
     @GetMapping
-    public ResponseEntity<ApiResponse<HouseListResponseDto>> getHouses(@PathVariable Long movingPlanId) {
+    public ResponseEntity<ApiResponse<HouseListResponseDto>> getHouses(@PathVariable Long movingPlanId, @AuthenticationPrincipal User user) {
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         return ResponseEntity.ok(
           ApiResponse.ok("OK", "집 목록을 조회에 성공했습니다.", houseService.getHouses(movingPlanId))
         );
     }
 
     @GetMapping("/{houseId}")
-    public ResponseEntity<ApiResponse<GetHouseDetailsResponseDto>> getHouseDetails(@PathVariable Long movingPlanId, @PathVariable Long houseId) {
+    public ResponseEntity<ApiResponse<GetHouseDetailsResponseDto>> getHouseDetails(@PathVariable Long movingPlanId, @PathVariable Long houseId, @AuthenticationPrincipal User user) {
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         return ResponseEntity.ok(
             ApiResponse.ok("OK", "집 조회에 성공했습니다.",  houseService.getHouseDetails(movingPlanId, houseId))
         );
     }
 
     @PatchMapping("/{houseId}/nickname")
-    public ResponseEntity<ApiResponse<ChangeHouseNicknameResponseDto>> updateHouseNickname(@PathVariable Long movingPlanId, @PathVariable Long houseId, @RequestBody @Valid ChangeHouseNicknameRequestDto requestDto) {
+    public ResponseEntity<ApiResponse<ChangeHouseNicknameResponseDto>> updateHouseNickname(@PathVariable Long movingPlanId, @PathVariable Long houseId, @RequestBody @Valid ChangeHouseNicknameRequestDto requestDto, @AuthenticationPrincipal User user) {
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         return ResponseEntity.ok(
             ApiResponse.ok("UPDATED", "집 별칭 수정에 성공했습니다.", houseService.updateHouseNickname(movingPlanId, houseId, requestDto))
         );
     }
 
     @PatchMapping("/{houseId}/content")
-    public ResponseEntity<ApiResponse<ChangeHouseContentResponseDto>> updateHouseContent(@PathVariable Long movingPlanId, @PathVariable Long houseId, @RequestBody @Valid ChangeHouseContentRequestDto requestDto) {
+    public ResponseEntity<ApiResponse<ChangeHouseContentResponseDto>> updateHouseContent(@PathVariable Long movingPlanId, @PathVariable Long houseId, @RequestBody @Valid ChangeHouseContentRequestDto requestDto, @AuthenticationPrincipal User user) {
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         return ResponseEntity.ok(
             ApiResponse.ok("UPDATED", "집 상세 내용 수정에 성공했습니다.", houseService.updateHouseContent(movingPlanId, houseId, requestDto))
         );
     }
     @DeleteMapping("/{houseId}")
-    public ResponseEntity<ApiResponse<Void>> deleteHouse(@PathVariable Long movingPlanId, @PathVariable Long houseId){
+    public ResponseEntity<ApiResponse<Void>> deleteHouse(@PathVariable Long movingPlanId, @PathVariable Long houseId, @AuthenticationPrincipal User user){
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         houseService.deleteHouse(movingPlanId, houseId);
 
         return ResponseEntity.ok(
