+++ BE/src/main/java/com/example/p24zip/domain/movingPlan/controller/MovingPlanController.java
@@ -5,6 +5,7 @@ import com.example.p24zip.domain.movingPlan.dto.response.MovingPlanResponseDto;
 import com.example.p24zip.domain.movingPlan.service.MovingPlanService;
 import com.example.p24zip.domain.user.entity.User;
 import com.example.p24zip.global.response.ApiResponse;
+import com.example.p24zip.global.validator.MovingPlanValidator;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -21,6 +22,7 @@ import java.util.Map;
 public class MovingPlanController {
 
     private final MovingPlanService movingPlanService;
+    private final MovingPlanValidator movingPlanValidator;
 
     @PostMapping
     public ResponseEntity<ApiResponse<MovingPlanResponseDto>> createMovingPlan(
@@ -55,6 +57,8 @@ public class MovingPlanController {
             @Valid @RequestBody MovingPlanRequestDto requestDto,
             @AuthenticationPrincipal User user) {
 
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
+
         return ResponseEntity.ok(ApiResponse.ok(
                 "UPDATED",
                 "플랜 제목 수정에 성공했습니다.",
@@ -66,6 +70,8 @@ public class MovingPlanController {
     public ResponseEntity<ApiResponse<Object>> deleteMovingPlan(
             @PathVariable Long movingPlanId,
             @AuthenticationPrincipal User user) {
+
+        movingPlanValidator.validateMovingPlanAccess(movingPlanId, user);
 
         movingPlanService.deleteMovingPlan(movingPlanId);
